Changed to use single threading and nonblocking I/O (including for SSL).
This required a significant amount of restructuring; e.g., had to move
the event loop from marinda-gs to GlobalSpace.  Also, quite a lot had
to change with connections and I/O to deal with nonblocking I/O.
Not everything is done; specifically, an SSL read that requires waiting
on write, and similarly for SSL writes waiting on reads, aren't handled.

Other minor improvements:

 * BasicSocket.do_not_reverse_lookup is now true.
 * marinda-gs removes the lock file after checkpoint-and-exit.