Changed to use configured server host name for validation of certs.
ClientSSLConnection#open now uses the server host name configured in
local-config.yaml for validating the server name in the server cert.
Before, the method used the hostname obtained with a reverse DNS
lookup (via BasicSocket.do_not_reverse_lookup = false).  The prior
behavior is likely more secure, but it requires enabling reverse
lookups on connections, which we prefer to disable.